Copy data from build_dir to releases/XXX

+++ hooks/post-receive.sh
@@ -4,7 +4,7 @@
 # Command to get two IDs for call this hook for testing
 # git log -2 --format=oneline --reverse
 # ./hooks/post-receive $FROM_ID $TO_ID master
-umask 002
+umask 022
 
 # Parameters
 if ! [ -t 0 ]; then
@@ -62,23 +62,25 @@ echo "Go to build the project"
 $HOME/bin/nodejs/bin/compile "$build_dir" "$cache_dir" "$config_dir"
 if [ $? -eq 0 ]; then
     echo "Deploy project: $repo_name"
-    mv $build_dir "$releases_dir/$newrev"
+    mkdir "$releases_dir/$newrev"
+    cp -rp "$build_dir/*" "$releases_dir/$newrev/"
+    rm -rf $build_dir
     echo "Stop server nginx"
     sudo service nginx stop
     echo "Stop forever"
     forever stop "$www_dir/current/server.js"
     cd $www_dir
     echo "Change the link current on the new release"
-    rm -f current && ln -s "./releases/$newrev" current
-    chmod -R g+w "$www_dir/current/"
-    sudo chown -R www-data:nogroup "$www_dir/current/"
+    if [ -f "$www_dir/current" ]; then rm -f current && ln -s "./releases/$newrev" current fi
+    #chmod -R g+w "$www_dir/current/"
+    sudo chown -R git:www-data "$www_dir/current/"
     echo "Launch forever"
     forever start -m 5 -p "$www_dir" -a -l "$log_dir/forever.log" -o "$log_dir/$repo_name.log" -e "$log_dir/error.log" --pidFile "$run_dir/$repo_name.pid" --sourceDir "$www_dir/current/" --minUptime 1000 --spinSleepTime 5000 server.js
     echo "Launch server nginx"
     sudo service nginx start
 else
-    echo >&2 "Error in compile script"
     rm -rf $build_dir
+    echo >&2 "Error in compile script"
     exit 1
 fi
